aula_07_0/requirements.txt
numpy
PyOpenGL
PyOpenGL-accelerate
pygame